requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
gpxpy
scipy
pandas
python-dotenv
slowapi